requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
docopt==0.6.2             # via -r requirements.in
numpy==1.19.3             # via -r requirements.in, opencv-python
opencv-python==4.4.0.44   # via -r requirements.in
pdf2image==1.14.0         # via -r requirements.in
pillow==8.0.1             # via -r requirements.in, pdf2image
